Refactor search conferences code
