+++ src/conf_world/lib/main.dart
@@ -1,7 +1,6 @@
+import 'package:conf_world/view/home_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'Navigation.dart';
-import 'MyButton.dart';
 
 /*
 void main() {
@@ -40,40 +39,3 @@ void main() {
   ));
 }
 
-class HomePageState extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-      title: Text("Conf World"),
-      centerTitle: true,
-      backgroundColor: Colors.green[500],
-    ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              FlatButton(
-                padding: EdgeInsets.zero,
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => Navigation(filter: 'false')),
-                  );
-                },
-                child: MyButton(
-                  name: 'See World Map',
-                ),
-              ),
-              FlatButton(
-                onPressed: null,
-                child: MyButton(
-                  name: 'Exit',
-                ),
-              ),
-            ],
-          ),
-        ));
-  }
-}
