+++ src/conf_world/lib/model/conference_model.dart
@@ -1,0 +1,11 @@
+class ConferenceInfo {
+  final String name;
+  final String type;
+  final String date;
+  final String description;
+
+  final double latitude;
+  final double longitude;
+
+  ConferenceInfo(this.name, this.type, this.date, this.description, this.latitude, this.longitude);
+}
